+++ lib/__tests__/find-config.test.js
@@ -13,7 +13,7 @@ describe('find-config', () => {
     const cosmiconfig = require('cosmiconfig')
     cosmiconfig.cosmiconfigSync.mockReturnValue({ search })
 
-    const findConfig = require('./find-config')
+    const findConfig = require('../find-config')
     const result = findConfig('babel')
 
     expect(cosmiconfig.cosmiconfigSync).toHaveBeenCalledWith('babel', {})
@@ -29,7 +29,7 @@ describe('find-config', () => {
     const cosmiconfig = require('cosmiconfig')
     cosmiconfig.cosmiconfigSync.mockReturnValue({ search })
 
-    const findConfig = require('./find-config')
+    const findConfig = require('../find-config')
     const result = findConfig('eslint')
 
     expect(cosmiconfig.cosmiconfigSync).toHaveBeenCalledWith('eslint', {})
@@ -46,7 +46,7 @@ describe('find-config', () => {
     const cosmiconfig = require('cosmiconfig')
     cosmiconfig.cosmiconfigSync.mockReturnValue({ load, search })
 
-    const findConfig = require('./find-config')
+    const findConfig = require('../find-config')
     const result = findConfig('jest', {}, 'path/to/jest.config.js')
 
     expect(cosmiconfig.cosmiconfigSync).toHaveBeenCalledWith('jest', {})
