+++ lib/__tests__/get-config-path.test.js
@@ -1,9 +1,9 @@
 'use strict'
 
-const getConfigPath = require('./get-config-path')
-const findConfig = require('./find-config')
+const getConfigPath = require('../get-config-path')
+const findConfig = require('../find-config')
 
-jest.mock('./find-config')
+jest.mock('../find-config')
 
 /** @type {Object} */
 const mockFindConfig = findConfig
