+++ lib/__tests__/get-config.test.js
@@ -1,9 +1,9 @@
 'use strict'
 
-const getConfig = require('./get-config')
-const findConfig = require('./find-config')
+const getConfig = require('../get-config')
+const findConfig = require('../find-config')
 
-jest.mock('./find-config')
+jest.mock('../find-config')
 
 /** @type {Object} */
 const mockFindConfig = findConfig
