+++ lib/__tests__/get-ignore-path.test.js
@@ -1,9 +1,9 @@
 'use strict'
 
-const getIgnorePath = require('./get-ignore-path')
-const findConfig = require('./find-config')
+const getIgnorePath = require('../get-ignore-path')
+const findConfig = require('../find-config')
 
-jest.mock('./find-config')
+jest.mock('../find-config')
 
 /** @type {Object} */
 const mockFindConfig = findConfig
