+++ lib/__tests__/index.test.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const index = require('.')
+const index = require('..')
 
 describe('index', () => {
   it('is defined', () => {
