+++ lib/__tests__/messages.test.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const messages = require('./messages')
+const messages = require('../messages')
 
 describe('messages', () => {
   it('is defined', () => {
