+++ lib/__tests__/parse-script.test.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const parseScript = require('./parse-script')
+const parseScript = require('../parse-script')
 
 describe('parse-script', () => {
   it.each([
