+++ lib/__tests__/parse-scripts.test.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const parseScripts = require('./parse-scripts')
+const parseScripts = require('../parse-scripts')
 
 const scriptMap = {
   a: 'path/to/a',
