+++ lib/__tests__/rewire-args.test.js
@@ -2,7 +2,7 @@
 
 /* eslint-disable max-params */
 
-const rewireArgs = require('./rewire-args')
+const rewireArgs = require('../rewire-args')
 
 test.each([
   // Should apply default options to args
