+++ lib/__tests__/rewire.test.js
@@ -1,9 +1,9 @@
 'use strict'
 
-const rewire = require('./rewire')
-const run = require('./run')
+const rewire = require('../rewire')
+const run = require('../run')
 
-jest.mock('./run')
+jest.mock('../run')
 
 const ARGV = process.argv
