+++ lib/__tests__/run-script.test.js
@@ -2,10 +2,10 @@
 
 /* eslint-disable max-lines-per-function */
 
-const runScript = require('./run-script')
-const run = require('./run')
+const runScript = require('../run-script')
+const run = require('../run')
 
-jest.mock('./run')
+jest.mock('../run')
 
 const mockExecaResult = (props) => ({
   command: null,
@@ -86,7 +86,9 @@ describe('run-script', () => {
     expect(run).toHaveBeenCalledWith(
       'node',
       expect.arrayContaining([script.path, ...script.args]),
-      expect.objectContaining({ env: script.options.env })
+      expect.objectContaining({
+        env: expect.objectContaining(script.options.env),
+      })
     )
   })
 
@@ -105,7 +107,9 @@ describe('run-script', () => {
     expect(run).toHaveBeenCalledWith(
       'node',
       expect.arrayContaining([script.path, ...script.args]),
-      expect.objectContaining({ env: script.options.env })
+      expect.objectContaining({
+        env: expect.objectContaining(script.options.env),
+      })
     )
   })
