+++ lib/__tests__/run-scripts.test.js
@@ -1,9 +1,9 @@
 'use strict'
 
-const runScript = require('./run-script')
-const runScripts = require('./run-scripts')
+const runScript = require('../run-script')
+const runScripts = require('../run-scripts')
 
-jest.mock('./run-script')
+jest.mock('../run-script')
 
 /** @type {Object} */
 const mockRunScript = runScript
