+++ lib/__tests__/run.test.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const execa = require('execa')
-const run = require('./run')
+const run = require('../run')
 
 jest.mock('execa')
