+++ lib/__tests__/wire.run-scripts.commands.test.js
@@ -1,9 +1,9 @@
 'use strict'
 
-const runScripts = require('./run-scripts')
-const wire = require('./wire')
+const runScripts = require('../run-scripts')
+const wire = require('../wire')
 
-jest.mock('./run-scripts')
+jest.mock('../run-scripts')
 
 /** @type {Object} */
 const mockRun = runScripts
