+++ lib/__tests__/wire.run.test.js
@@ -2,10 +2,10 @@
 
 /* eslint-disable max-lines-per-function */
 
-const run = require('./run')
-const wire = require('./wire')
+const run = require('../run')
+const wire = require('../wire')
 
-jest.mock('./run')
+jest.mock('../run')
 
 /** @type {Object} */
 const mockRun = run
